S402 t3128 add map to destination create

See merge request seng302-2019/team-300!150

+++ backend/app/controllers/dto/Destination/CreateDestReq.java
@@ -22,7 +22,7 @@ public class CreateDestReq {
     @Constraints.Required
     public boolean isDeleted;
 
-    public List<Integer> travellerTypes;
+    public List<Long> travellerTypes;
 
     public String getName() {
         return name;
